Full CRUD achieved

+++ backend/controllers/musicController.js
@@ -1,12 +1,13 @@
 const asyncHandler = require('express-async-handler')
 
 const Music = require('../models/musicModel')
+const User = require('../models/userModel')
 
 // @desc Get music
 // @route GET /api/music
 // @access Private
 const getMusic = asyncHandler(async (req, res) => {
-    const music = await Music.find()
+    const music = await Music.find({ user: req.user.id })
     
     res.status(200).json(music)
 })
@@ -22,6 +23,7 @@ const setMusic = asyncHandler(async (req, res) => {
 
     const music = await Music.create({
         text: req.body.text,
+        user: req.user.id,
     })
 
     res.status(200).json(music)
@@ -36,6 +38,20 @@ const updateMusic = asyncHandler(async (req, res) => {
     if(!music) {
         res.status(400)
         throw new Error('Music not found')
+    }
+
+    const user = await User.findById(req.user.id)
+ 
+    // Check for user
+    if(!user) {
+        res.status(401)
+        throw new Error('User not found')
+    }
+
+    // Make sure logged in user matches the music user
+    if(music.user.toString() !== user.id) {
+        res.status(401)
+        throw new Error('User not authorized')
     }
 
     const updatedMusic = await Music.findByIdAndUpdate(req.params.id, req.
@@ -55,6 +71,20 @@ const deleteMusic = asyncHandler(async (req, res) => {
     if(!music) {
         res.status(400)
         throw new Error('Music not found')
+    }
+
+    const user = await User.findById(req.user.id)
+ 
+    // Check for user
+    if(!user) {
+        res.status(401)
+        throw new Error('User not found')
+    }
+
+    // Make sure logged in user matches the music user
+    if(music.user.toString() !== user.id) {
+        res.status(401)
+        throw new Error('User not authorized')
     }
 
     await Music.findByIdAndRemove(req.params.id);
